flask-server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
scikit-learn==1.1.3
python-dotenv==1.0.0
passlib==1.7.4